Huge overhaul of how modules work
Each module has its own folder within modules/ within each folder there is a file that contains all module information for that module. This is called moduleInfo.php

Within that you can set hooks that are picked up throughot the software, currently the only hooks are for the menus.

With this modifiction this will help make GLV2 alot more modular, when you want to install a 3rd party module it should be plug in and play with no code alterations.